Separate NotFound page and add InternalError page
